Auto approve scopes
This is to avoid the authorization page in
UAA. It is not clear how to remove approvals therefore
once they are approved the user has the scopes approved
until they expire. If the user has some approved scopes,
they will not be prompted to approve them until they
expire hence the ui flow differs when the user has to
approve scopes and when not.